refactor: reorganize modules into packages
